Replaced pseudo phys rm with allocator and repository
Apparently, Genode does not like attaching region maps to other region maps where offset != 0. Due to this reason, had to get rid of pseudo phys rm the purpose of which was to give addresses to ram dataspaces and keeping the mappings. Replaced it with AVL allocator and repository containing objects with ram dataspace, pseudo address and size